refactor: candidate resolution leaves ChunkPlan (#11)
ChunkCandidates holds the first pass of planning: is this chunk city,
does a road or highway or railway run through it, is it one section of a
multi-chunk building, and which building would go here. The answer is a
ChunkCandidate, cached per coordinate, and it is deliberately upstream of
everything a ChunkPlan decides - a candidate may not read a building
decision, its own or a neighbour's, or the decision graph stops being
acyclic. That constraint now has a class boundary rather than a comment.

calculateTopLeft stayed behind: it looked dead from inside the moved
block, and is not - the constructor calls it, and the constructor is not
moving.

CityStyleLookupSitesTest needed updating, which is the test doing its
job. It pins the exact set of files that resolve a city style by name so
that a new lookup site fails the build; this is the same lookup in a new
file rather than a new one, confirmed by checking that ChunkPlan went
from one occurrence to zero as ChunkCandidates went from zero to one. Its
javadoc warns against widening the list, so the entry moved rather than
being added to.

Pure code motion, verified token-for-token: 1455 tokens on both sides.

ChunkPlan 1193 -> 965 lines, from 1939 - under half.

All seven digest suites unchanged.
